refactor: moved test objects to separate project, only used from tests
