+++ AdSecGH/Components/3_Section/FlattenRebar.cs
@@ -1,41 +1,61 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Drawing;
 using System.Linq;
 
 using AdSecGH.Helpers;
 using AdSecGH.Helpers.GH;
 using AdSecGH.Parameters;
+using AdSecGH.Properties;
 
 using Grasshopper.Kernel;
 
 using Oasys.AdSec;
 using Oasys.AdSec.Reinforcement.Groups;
 using Oasys.AdSec.Reinforcement.Preloads;
+using Oasys.Business;
+using Oasys.GH.Helpers;
 using Oasys.Profiles;
 
 using OasysGH;
-using OasysGH.Components;
 using OasysGH.Parameters;
 using OasysGH.Units;
 
 using OasysUnits;
 
-namespace AdSecGH.Components {
-  public class FlattenRebar : GH_OasysComponent {
-    public override Guid ComponentGuid => new Guid("879ecac6-464d-4286-9113-c15fcf03e4e6");
-    public override GH_Exposure Exposure => GH_Exposure.tertiary | GH_Exposure.obscure;
-    public override OasysPluginInfo PluginInfo => AdSecGH.PluginInfo.Instance;
-    protected override System.Drawing.Bitmap Icon => Properties.Resources.FlattenRebar;
+using Attribute = Oasys.Business.Attribute;
 
-    public FlattenRebar() : base(
-      "FlattenRebars",
-      "FRb",
-      "Flatten all rebars in a section into single bars.",
-      CategoryName.Name(),
-      SubCategoryName.Cat4()) {
+namespace AdSecGH.Components {
+  public class FlattenRebarComponent : IBusinessComponent {
+
+    public ComponentAttribute Metadata { get; set; } = new ComponentAttribute {
+      Name = "FlattenRebar",
+      NickName = "FRb",
+      Description = "Flatten all rebars in a section into single bars.",
+    };
+    public ComponentOrganisation Organisation { get; set; } = new ComponentOrganisation {
+      Category = CategoryName.Name(),
+      SubCategory = SubCategoryName.Cat4(),
+    };
+    public Attribute[] GetAllInputAttributes() { throw new NotImplementedException(); }
+
+    public Attribute[] GetAllOutputAttributes() { throw new NotImplementedException(); }
+
+    public void UpdateInputValues(params object[] values) { throw new NotImplementedException(); }
+
+    public void Compute() { throw new NotImplementedException(); }
+  }
+
+  public class FlattenRebar : BusinessOasysGlue<FlattenRebarComponent> {
+
+    public FlattenRebar() {
       Hidden = true;
     }
 
+    public override Guid ComponentGuid => new Guid("879ecac6-464d-4286-9113-c15fcf03e4e6");
+    public override GH_Exposure Exposure => GH_Exposure.tertiary | GH_Exposure.obscure;
+    public override OasysPluginInfo PluginInfo => AdSecGH.PluginInfo.Instance;
+    protected override Bitmap Icon => Resources.FlattenRebar;
 
     protected override void RegisterInputParams(GH_InputParamManager pManager) {
       pManager.AddGenericParameter("Section", "Sec", "AdSec Section to get single rebars from", GH_ParamAccess.item);
@@ -50,17 +70,20 @@ namespace AdSecGH.Components {
       IQuantity force = new Force(0, DefaultUnits.ForceUnit);
       string forceUnitAbbreviation = string.Concat(force.ToString().Where(char.IsLetter));
 
-      pManager.AddGenericParameter("Position [" + lengthUnitAbbreviation + "]", "Vx", "Rebar position as 2D vertex in the section's local yz-plane ", GH_ParamAccess.list);
+      pManager.AddGenericParameter("Position [" + lengthUnitAbbreviation + "]", "Vx",
+        "Rebar position as 2D vertex in the section's local yz-plane ", GH_ParamAccess.list);
       pManager.HideParameter(0);
-      pManager.AddGenericParameter("Diameter [" + lengthUnitAbbreviation + "]", "Ø", "Bar Diameter", GH_ParamAccess.list);
+      pManager.AddGenericParameter("Diameter [" + lengthUnitAbbreviation + "]", "Ø", "Bar Diameter",
+        GH_ParamAccess.list);
       pManager.AddIntegerParameter("Bundle Count", "N", "Count per bundle (1, 2, 3 or 4)", GH_ParamAccess.list);
-      pManager.AddGenericParameter("PreLoad", "P", "The pre-load per reinforcement bar. Positive value is tension.", GH_ParamAccess.list);
+      pManager.AddGenericParameter("PreLoad", "P", "The pre-load per reinforcement bar. Positive value is tension.",
+        GH_ParamAccess.list);
       pManager.AddGenericParameter("Material", "Mat", "Material Type", GH_ParamAccess.list);
     }
 
     protected override void SolveInstance(IGH_DataAccess DA) {
       // get section input
-      AdSecSection section = AdSecInput.AdSecSection(this, DA, 0);
+      var section = AdSecInput.AdSecSection(this, DA, 0);
 
       // create flattened section
       ISection flat = null;
@@ -79,11 +102,11 @@ namespace AdSecGH.Components {
       var materialType = new List<string>();
 
       // loop through rebar groups in flattened section
-      foreach (IGroup rebargrp in flat.ReinforcementGroups) {
+      foreach (var rebargrp in flat.ReinforcementGroups) {
         try // first try if not a link group type
         {
           var snglBrs = (ISingleBars)rebargrp;
-          foreach (IPoint pos in snglBrs.Positions) {
+          foreach (var pos in snglBrs.Positions) {
             // position
             pointGoos.Add(new AdSecPointGoo(pos));
