+++ AdSecGH/Helpers/BusinessExtensions.cs
@@ -88,69 +88,4 @@ namespace Oasys.GH.Helpers {
     }
   }
 
-  public class DummyBusiness : IBusinessComponent {
-
-    public DoubleParameter Alpha { get; set; } = new DoubleParameter {
-      Name = "Alpha",
-      NickName = "A",
-      Description = "Alpha description",
-      Default = 1,
-    };
-    public DoubleParameter Beta { get; set; } = new DoubleParameter {
-      Name = "Beta",
-      NickName = "B",
-      Description = "Beta description",
-      Default = 2,
-    };
-
-    public ComponentAttribute Metadata { get; set; } = new ComponentAttribute {
-      Name = "Dummy Business",
-      NickName = "DB",
-      Description = "Dummy Business Description",
-    };
-    public ComponentOrganisation Organisation { get; set; } = new ComponentOrganisation {
-      Category = "Dummy Category",
-      SubCategory = "Dummy SubCategory",
-    };
-
-    public Attribute[] GetAllInputAttributes() {
-      return new[] {
-        Alpha,
-      };
-    }
-
-    public Attribute[] GetAllOutputAttributes() {
-      return new[] {
-        Beta,
-      };
-    }
-
-    public void UpdateInputValues(params object[] values) { }
-
-    public void Compute() { Beta.Value = (Alpha.Value * 2) + 10; }
-
-    public void GetDefaultValues() { throw new NotImplementedException(); }
-  }
-
-  public static class Test {
-
-    public const string thisIsForTestingPurposesOnly = "this is for testing purposes only";
-  }
-
-  public class DummyOasysComponent : BusinessOasysGlue<DummyBusiness> {
-
-    public DummyOasysComponent() : base("Oasys Component Glue", "OCG", Test.thisIsForTestingPurposesOnly, "Oasys",
-      "dummy") { }
-
-    public override GH_Exposure Exposure { get; } = GH_Exposure.hidden;
-
-    public override Guid ComponentGuid => new Guid("CAA08C9E-417C-42AE-B734-91F214C8B87F");
-
-    protected override void SolveInstance(IGH_DataAccess DA) { }
-  }
-
-  public class Dummy : BusinessGlue<DummyBusiness> {
-    public override GH_Exposure Exposure { get; } = GH_Exposure.hidden;
-    public override Guid ComponentGuid => new Guid("CAA08C9E-417C-42AE-B704-91F214C8C871");
-  }
 }
