+++ AdSecGHTests/Helpers/DropIntoGhDocumentTest.cs
@@ -3,8 +3,6 @@ using System.Threading.Tasks;
 
 using Grasshopper.Kernel;
 
-using Oasys.GH.Helpers;
-
 using Xunit;
 
 namespace AdSecGHTests.Helpers {
